requirements.txt
langchain==0.0.215
openai==0.27.8
chromadb==0.3.26 
python-docx==0.8.11
tiktoken==0.4.0
#unstructured==0.7.8